Update Button.swift
When using libraries like Bootstrap, I often need to put child elements inside buttons. But the init() only takes contents.

Not sure if it was intentional. To only allow users to create Buttons with String contents. But I would think all Tags should have a designated init() that allows for children with a builder parameter, whether or not it produces well written HTML. And then the convenience inits would implement preferred or common usage. Like in A.swift where the convenience init() just takes contents.